requirements.txt
bs4~=0.0.1
beautifulsoup4~=4.9.3
requests~=2.25.1
python-dotenv~=0.15.0
SQLAlchemy~=1.3.23